move py files to folder
