+++ Xcao19_yjhang_zy0105/join_by_ZIP.py
@@ -7,6 +7,12 @@ import dml
 import prov.model
 import datetime
 import uuid
+import math
+def dealWithZip(data,key):
+    for i in data:
+        if(len(str(int(i[key][:5]))) == 4):
+            i[key] = '0'+str(int(i[key][:5]))
+
 
 
 class join_by_ZIP(dml.Algorithm):
@@ -40,7 +46,15 @@ class join_by_ZIP(dml.Algorithm):
         policeStation = list(r_policeStation.find({},{'_id':0,'ZIP':1}))
         property = list(r_property.find({},{'_id':0,'ZIPCODE':1,'AV_TOTAL':1,'GROSS_AREA':1}))
         school = list(r_school.find({},{'_id':0,'ZIP':1}))
-        #print(center)
+        dealWithZip(center,"ZIP")
+        dealWithZip(centerPool,"ZIP")
+        dealWithZip(policeStation,"ZIP")
+        #dealWithProZip(property,"ZIPCODE")
+        dealWithZip(school,"ZIP")
+        for s in property:
+            s["ZIPCODE"] = "0"+ str(s["ZIPCODE"])[:4]
+        # print(school)
+
         val_dic = {}
         ar_dic = {}
         ppty_avg = {}
@@ -48,13 +62,13 @@ class join_by_ZIP(dml.Algorithm):
         for ppty in property:
             #val_dic.setdefault(ppty['ZIPCODE'],0)
             if(ppty['GROSS_AREA'] is not None):
-                if(val_dic.__contains__(str(ppty['ZIPCODE'])[:-2])):
-                    val_dic[str(ppty['ZIPCODE'])[:-2]]+=ppty['AV_TOTAL']
+                if(val_dic.__contains__(ppty['ZIPCODE'])):
+                    val_dic[ppty['ZIPCODE']]+=ppty['AV_TOTAL']
                     #print(ppty['GROSS_AREA'])
-                    ar_dic[str(ppty['ZIPCODE'])[:-2]]+=ppty['GROSS_AREA']
+                    ar_dic[ppty['ZIPCODE']]+=ppty['GROSS_AREA']
                 else:
-                    val_dic[str(ppty['ZIPCODE'])[:-2]]=ppty['AV_TOTAL']
-                    ar_dic[str(ppty['ZIPCODE'])[:-2]]=ppty['GROSS_AREA']
+                    val_dic[ppty['ZIPCODE']]=ppty['AV_TOTAL']
+                    ar_dic[ppty['ZIPCODE']]=ppty['GROSS_AREA']
         #print(val_dic)
 
         for k in val_dic:
